+++ Lti1p3Bundle.php
@@ -27,6 +27,7 @@ use OAT\Bundle\Lti1p3Bundle\DependencyInjection\Security\Factory\Message\LtiPlat
 use OAT\Bundle\Lti1p3Bundle\DependencyInjection\Security\Factory\Message\LtiToolMessageSecurityFactory;
 use OAT\Bundle\Lti1p3Bundle\DependencyInjection\Security\Factory\Service\LtiServiceSecurityFactory;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class Lti1p3Bundle extends Bundle
@@ -38,8 +39,12 @@ class Lti1p3Bundle extends Bundle
         // Bundle dependencies builder pass
         $container->addCompilerPass(new BuilderPass());
 
-        $securityExtension = $container->getExtension('security');
+        // Bundle security extension configuration
+        $this->configureSecurityExtension($container->getExtension('security'));
+    }
 
+    private function configureSecurityExtension(ExtensionInterface $securityExtension): void
+    {
         // LTI platform messages security registration
         $securityExtension->addSecurityListenerFactory(new LtiPlatformMessageSecurityFactory());